Removed unnecessary files and code that was previously used for GraphQL backend in order to prepare for REST api migration
