+++ Nutrivurv/Nutrivurv/LSLDashboardViewController.swift
@@ -48,49 +48,6 @@ class LSLDashboardViewController: UIViewController {
         viewController.modalTransitionStyle = .flipHorizontal
         self.present(viewController, animated: true, completion: nil)
     }
-    
-    // MARK: - Profile Laod and Update
-    
-//    private func checkForProfile() {
-//        dashboardController.checkForProfile { (hasProfile) in
-//            if hasProfile {
-//                DispatchQueue.main.async {
-//                    self.loadProfile()
-//                }
-//            } else {
-//                let destination = UIStoryboard(name: "Profile", bundle: .main)
-//                guard let profileCreationVC = destination.instantiateInitialViewController() as? LSLCalculateBMIViewController else {
-//                    print("Unable to instantiate profile creation view controller")
-//                    return
-//                }
-//                profileCreationVC.createProfileDelegate = self
-//                profileCreationVC.nutritionController = self.userController
-//                self.navigationController?.pushViewController(profileCreationVC, animated: true)
-//            }
-//        }
-//    }
-    
-//    private func loadProfile() {
-//        self.dashboardController.getMyName { (result) in
-//            if let name = try? result.get() {
-//                self.navigationItem.title = name
-//            } else {
-//                print("Couldn't get name: \(result)")
-//            }
-//        }
-//
-//        self.dashboardController.getMyWeight { (result) in
-//            if let weight = try? result.get() {
-//                self.currentWeightLabel.text = String(weight)
-//            } else {
-//                if let weight = LSLUserController.weight {
-//                    self.currentWeightLabel.text = String(weight)
-//                } else {
-//                    print("Unable to load weight for user")
-//                }
-//            }
-//        }
-//    }
 }
 
 // MARK: - Profile Completion Protocol Declaration & Delegate Conformance
