--- Nutrivurv/Nutrivurv/Network.swift
@@ -1,185 +1,0 @@
-//
-//  Network.swift
-//  Nutrition Tracker
-//
-//  Created by Michael Stoffer on 3/16/20.
-//  Copyright © 2020 Lambda School. All rights reserved.
-//
-
-import Foundation
-import Apollo
-import KeychainSwift
-
-enum NetworkError: Error {
-    case noAuth
-    case badAuth
-    case otherError
-    case badData
-    case noEncode
-    case noDecode
-}
-
-class Network {
-    static let shared = Network()
-    static let loginKeychainKey = "login"
-    var apollo: ApolloClient = createApolloClient()
-    
-    private init() { }
-    
-    static func createApolloClient() -> ApolloClient {
-        let keychain = KeychainSwift()
-        let token = keychain.get(Network.loginKeychainKey) ?? ""
-        let url = URL(string: "https://protected-forest-13208.herokuapp.com/")!
-        let configuration = URLSessionConfiguration.default
-        
-        if token != "" {
-            configuration.httpAdditionalHeaders = ["authorization": "Bearer \(token)"]
-        }
-
-        return ApolloClient(
-            networkTransport: HTTPNetworkTransport(url: url, session: URLSession(configuration: configuration))
-        )
-    }
-    
-    func loginUser(email: String, password: String, completion: @escaping (Result<Bool, NetworkError>) -> Void) {
-        apollo.perform(mutation: LoginMutation(data: LoginUserInput(email: email, password: password))) { result in
-            switch result {
-                case .success(let graphQLResult):
-                    guard let token = graphQLResult.data?.login.token else {
-                        print("Bad Data: \(graphQLResult)")
-                        DispatchQueue.main.async {
-                            completion(.failure(.badAuth))
-                        }
-                        return
-                    }
-                    
-                    guard graphQLResult.errors == nil else {
-                        print("Errors from server: \(graphQLResult.errors!)")
-                        DispatchQueue.main.async {
-                            completion(.failure(.otherError))
-                        }
-                        return
-                    }
-                    
-                    let keychain = KeychainSwift()
-                    keychain.set(token, forKey: Network.loginKeychainKey)
-                    self.apollo = Network.createApolloClient()
-
-                    DispatchQueue.main.async {
-                        completion(.success(true))
-                    }
-                case .failure(let error):
-                    print("Error: \(error)")
-                    DispatchQueue.main.async {
-                        completion(.failure(.noAuth))
-                    }
-            }
-        }
-    }
-    
-    func createUser(name: String, email: String, password: String, completion: @escaping (Result<Bool, NetworkError>) -> Void) {
-        self.apollo.perform(mutation: CreateUserMutation(data: CreateUserInput(name: name, email: email, password: password))) { result in
-            switch result {
-            case .success(let graphQLResult):
-                guard let token = graphQLResult.data?.createUser.token else {
-                    // If an account already exists with the same credentials, this completion will run
-                    DispatchQueue.main.async {
-                        completion(.failure(.badAuth))
-                    }
-                    return
-                }
-                
-                guard graphQLResult.errors == nil else {
-                    print("Errors from server: \(graphQLResult.errors!)")
-                    DispatchQueue.main.async {
-                        completion(.failure(.otherError))
-                    }
-                    return
-                }
-                
-                let keychain = KeychainSwift()
-                keychain.set(token, forKey: Network.loginKeychainKey)
-                self.apollo = Network.createApolloClient()
-                
-                DispatchQueue.main.async {
-                    completion(.success(true))
-                }
-                
-            case .failure(let error):
-                print("Error: \(error)")
-                DispatchQueue.main.async {
-                    completion(.failure(.noAuth))
-                }
-            }
-        }
-    }
-    
-    func createProfile(age: Int, weight: Int, height: Int, gender: Bool?, goalWeight: Int?, activityLevel: Int?, diet: String?, completion: @escaping (Result<Bool, NetworkError>) -> Void) {
-        apollo.perform(mutation: CreateProfileMutation(data: CreateProfileInput(age: age, weight: weight, height: height, gender: gender, goalWeight: goalWeight, activityLevel: activityLevel, diet: diet))) { result in
-            switch result {
-            case .success(let graphQLResult):
-                guard graphQLResult.data?.createProfile.id != nil else {
-                    DispatchQueue.main.async {
-                        completion(.failure(.badData))
-                    }
-                    return
-                }
-                
-                guard graphQLResult.errors == nil else {
-                    print("Errors from server: \(graphQLResult.errors!)")
-                    DispatchQueue.main.async {
-                        completion(.failure(.otherError))
-                    }
-                    return
-                }
-                
-                DispatchQueue.main.async {
-                    completion(.success(true))
-                }
-                
-            case .failure(let error):
-                print("Error: \(error)")
-                DispatchQueue.main.async {
-                    completion(.failure(.noAuth))
-                }
-            }
-        }
-    }
-    
-    static func isLoggedIn() -> Bool {
-      let keychain = KeychainSwift()
-        return keychain.get(Network.loginKeychainKey) != nil
-    }
-    
-    func createFoodLog(calories: Int, fat: Int, carbs: Int, fiber: Int, protein: Int, foodString: String, quantity: Int, mealType: String, completion: @escaping (Result<Bool, NetworkError>) -> Void) {
-        let currentDateTime = Date()
-        apollo.perform(mutation: CreateDailyRecordMutation(data: CreateDailyRecordInput(date: "\(currentDateTime)", calories: calories, fat: fat, carbs: carbs, fiber: fiber, protein: protein, foodString: foodString, quantity: quantity, mealType: mealType))) { result in
-            switch result {
-                case .success(let graphQLResult):
-                    guard graphQLResult.data?.createDailyRecord.createdAt != nil else {
-                        DispatchQueue.main.async {
-                            completion(.failure(.badData))
-                        }
-                        return
-                    }
-
-                    guard graphQLResult.errors == nil else {
-                        print("Errors from server: \(graphQLResult.errors!)")
-                        DispatchQueue.main.async {
-                            completion(.failure(.otherError))
-                        }
-                        return
-                    }
-
-                    DispatchQueue.main.async {
-                        completion(.success(true))
-                    }
-                case .failure(let error):
-                    print("Error: \(error)")
-                    DispatchQueue.main.async {
-                        completion(.failure(.noAuth))
-                    }
-            }
-        }
-    }
-}
