+++ Nutrivurv/Nutrivurv/NetworkError.swift
@@ -1,0 +1,18 @@
+//
+//  NetworkError.swift
+//  Nutrition Tracker
+//
+//  Created by Dillon on 6/23/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case noAuth
+    case badAuth
+    case otherError
+    case badData
+    case noEncode
+    case noDecode
+}
